Refactor about component
