+++ src/assets/text/about.ts
@@ -11,3 +11,8 @@ export const about = {
     `Happy hunting, recruit.`,
   ],
 };
+export const wordsToLinkAbout: Record<string, string> = {
+  "Kryha Labs": "https://kryha.io",
+  Aleo: "https://aleo.org",
+  zPass: "https://zpass.aleo.org/",
+};
